+++ src/mc_tmpl_8.rs
@@ -2900,8 +2900,8 @@ unsafe extern "C" fn put_bilin_c(
         src_stride as usize,
         w as usize,
         h as usize,
-        mx,
-        my,
+        mx as usize,
+        my as usize,
         BitDepth8::new(()),
     )
 }
